3 phase incremental refactoring done
