+++ client/src/components/payment/StudentInfoCard.jsx
@@ -1,535 +1,65 @@
-// File: client/src/components/PaymentModal.jsx
-// PaymentModal - PHASE 2 REFACTORED
-// Lines reduced from 650 to 420 (230 lines extracted in Phase 2)
-// Total reduction: 717 → 420 lines (41% reduction from original)
-// PRESERVED: All existing functionality - zero breaking changes
+// File: client/src/components/payment/StudentInfoCard.jsx
+// Lines 1-70: Student information display card with pricing breakdown
+// Extracted from PaymentModal.jsx (Lines 398-435)
+// Shows student details and their personalized pricing rates
 
-import { useEffect, useState, useRef } from "react";
-import { useAuth } from "../context/AuthContext";
-
-// Line 8-12: Phase 1 imports
-import PricingTierBadge from "./payment/PricingTierBadge";
-import { formatDate, getDisplayDate, getMaxDate, getMinDate } from "../utils/paymentDateUtils";
-
-// Line 14-17: Phase 2 NEW IMPORTS - extracted business logic
-import useStudentPricing from "../hooks/useStudentPricing";
-import usePaymentValidation from "../hooks/usePaymentValidation";
-import StudentInfoCard from "./payment/StudentInfoCard";
-
-const PaymentModal = ({
-  isOpen,
-  onClose,
-  student = null,
-  onPaymentSuccess,
-}) => {
-  const { token } = useAuth();
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState(null);
-  const [successMessage, setSuccessMessage] = useState(null);
-  
-  // Line 32-35: Phase 2 - Use custom pricing hook instead of manual state
-  const { 
-    studentPricing, 
-    pricingLoading, 
-    pricingError, 
-    getMembershipPrices 
-  } = useStudentPricing(student?.id, token, isOpen);
-  
-  // Line 40-42: Submission tracking to prevent duplicates
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const lastSubmissionRef = useRef(null);
-
-  const [formData, setFormData] = useState({
-    amount: "",
-    method: "CASH",
-    description: "",
-    extendMembership: true,
-    membershipType: "MONTHLY",
-    paymentDateOption: "today",
-    customPaymentDate: "",
-  });
-
-  const paymentMethods = [
-    { value: "CASH", label: "💵 Cash" },
-    { value: "CARD", label: "💳 Card" },
-    { value: "BANK_TRANSFER", label: "🏦 Bank Transfer" },
-    { value: "ONLINE", label: "🌐 Online Payment" },
-    { value: "CHECK", label: "📝 Check" },
-    { value: "OTHER", label: "📋 Other" },
-  ];
-
-  // Line 65-68: Phase 2 - Use validation hook
-  const { validateForm } = usePaymentValidation(formData, student, studentPricing);
-
-  // Line 71-90: Auto-update amount when pricing loads or membership type changes
-  useEffect(() => {
-    if (studentPricing) {
-      const prices = getMembershipPrices();
-      setFormData(prev => ({
-        ...prev,
-        amount: prices[prev.membershipType].toString()
-      }));
-    }
-  }, [studentPricing, getMembershipPrices]);
-
-  // Line 92-103: Handle membership type change
-  const handleMembershipTypeChange = (type) => {
-    const prices = getMembershipPrices();
-    
-    setFormData((prev) => ({
-      ...prev,
-      membershipType: type,
-      amount: prices[type].toString(),
-    }));
-
-    if (error && error.includes("amount")) {
-      setError(null);
-    }
-  };
-
-  // Line 106-123: Check for duplicate submission
-  const isDuplicateSubmission = (paymentData) => {
-    if (!lastSubmissionRef.current) return false;
-    
-    const last = lastSubmissionRef.current;
-    const now = Date.now();
-    
-    if (now - last.timestamp < 3000) {
-      if (
-        last.studentId === paymentData.studentId &&
-        last.amount === paymentData.amount &&
-        last.method === paymentData.method
-      ) {
-        console.warn("Duplicate submission blocked");
-        return true;
-      }
-    }
-    
-    return false;
-  };
-
-  // Line 126-203: Submit payment handler
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (isSubmitting) {
-      return;
-    }
-
-    // Line 135: Use extracted validation hook
-    if (!validateForm(setError)) {
-      return;
-    }
-
-    setIsSubmitting(true);
-    setLoading(true);
-    setError(null);
-    setSuccessMessage(null);
-
-    try {
-      let paymentDateToSend = null;
-      if (formData.paymentDateOption === "custom" && formData.customPaymentDate) {
-        paymentDateToSend = formData.customPaymentDate;
-      }
-
-      const tierContext = studentPricing?.isLegacy ? ` (${studentPricing.tier})` : '';
-      const paymentData = {
-        studentId: student.id,
-        amount: parseFloat(formData.amount),
-        method: formData.method,
-        description: formData.description || `${formData.membershipType} membership payment${tierContext}`,
-        extendMembership: formData.extendMembership,
-        membershipType: formData.membershipType,
-        ...(paymentDateToSend && { paymentDate: paymentDateToSend }),
-      };
-
-      if (isDuplicateSubmission(paymentData)) {
-        setError("Duplicate payment detected. Please wait before submitting again.");
-        return;
-      }
-
-      const response = await fetch("/api/payments/create", {
-        method: "POST",
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(paymentData),
-      });
-
-      const responseData = await response.json();
-
-      if (!response.ok) {
-        if (response.status === 409) {
-          throw new Error("Duplicate payment detected by server");
-        }
-        throw new Error(responseData.error || `Payment failed`);
-      }
-
-      lastSubmissionRef.current = {
-        timestamp: Date.now(),
-        studentId: paymentData.studentId,
-        amount: paymentData.amount,
-        method: paymentData.method
-      };
-
-      setSuccessMessage("Payment recorded successfully");
-
-      if (typeof onPaymentSuccess === 'function') {
-        onPaymentSuccess({
-          payment: responseData,
-          student: student,
-          studentPricing: responseData.studentPricing
-        });
-      }
-      
-      setTimeout(() => {
-        handleClose();
-      }, 1500);
-    } catch (error) {
-      console.error("Payment error:", error);
-      setError(error.message || "Payment processing failed");
-    } finally {
-      setLoading(false);
-      setTimeout(() => {
-        setIsSubmitting(false);
-      }, 2000);
-    }
-  };
-
-  // Line 206-228: Input handlers
-  const handleInputChange = (e) => {
-    const { name, value, type, checked } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: type === "checkbox" ? checked : value,
-    }));
-
-    if (error) {
-      setError(null);
-    }
-  };
-
-  const handlePaymentDateOptionsChange = (option) => {
-    setFormData((prev) => ({
-      ...prev,
-      paymentDateOption: option,
-      customPaymentDate: option === "today" ? "" : formatDate(new Date()),
-    }));
-
-    if (error) {
-      setError(null);
-    }
-  };
-
-  const handleCustomDateChange = (e) => {
-    setFormData((prev) => ({
-      ...prev,
-      customPaymentDate: e.target.value,
-    }));
-
-    if (error) {
-      setError(null);
-    }
-  };
-
-  // Line 231-248: Reset and close modal
-  const handleClose = () => {
-    setFormData({
-      amount: "",
-      method: "CASH",
-      description: "",
-      extendMembership: true,
-      membershipType: "MONTHLY",
-      paymentDateOption: "today",
-      customPaymentDate: "",
-    });
-    setError(null);
-    setSuccessMessage(null);
-    setIsSubmitting(false);
-    lastSubmissionRef.current = null;
-    onClose();
-  };
-
-  // Line 251-263: Escape key handler
-  useEffect(() => {
-    const handleKeyDown = (e) => {
-      if (e.key === "Escape" && isOpen && !loading) {
-        handleClose();
-      }
-    };
-
-    if (isOpen) {
-      document.addEventListener("keydown", handleKeyDown);
-      return () => document.removeEventListener("keydown", handleKeyDown);
-    }
-  }, [isOpen, loading]);
-
-  if (!isOpen) return null;
-
-  const membershipPrices = getMembershipPrices();
+/**
+ * StudentInfoCard Component
+ * Line 10-15: Displays student name, email, and pricing tier breakdown
+ * Shows monthly and yearly rates with grandfathered pricing indicator
+ * Only renders when student data is available
+ * 
+ * @param {Object} props
+ * @param {Object} props.student - Student object with name and email
+ * @param {Object} props.studentPricing - Pricing object with monthly/yearly rates
+ * @param {boolean} props.pricingLoading - Whether pricing is being fetched
+ * @returns {JSX.Element|null} Student info card or null if no student
+ */
+const StudentInfoCard = ({ student, studentPricing, pricingLoading }) => {
+  // Line 20-22: Don't render if no student selected
+  if (!student) return null;
 
   return (
-    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-      <div className="bg-white rounded-lg p-6 w-full max-w-md mx-4 max-h-[90vh] overflow-y-auto shadow-xl relative">
-        {/* Line 275-295: Header with extracted PricingTierBadge */}
-        <div className="flex justify-between items-center mb-4">
-          <div>
-            <h2 className="text-xl font-semibold text-gray-900">💳 Record Payment</h2>
-            {pricingLoading && (
-              <div className="mt-1 text-sm text-blue-600">Loading pricing...</div>
-            )}
-            {studentPricing && !pricingLoading && (
-              <div className="mt-2">
-                <PricingTierBadge pricing={studentPricing} />
-              </div>
-            )}
-            {pricingError && (
-              <div className="mt-1 text-sm text-orange-600">{pricingError}</div>
-            )}
-          </div>
-          <button
-            onClick={handleClose}
-            className="text-gray-400 hover:text-gray-600 transition-colors"
-            disabled={loading || isSubmitting}
-          >
-            <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-
-        {/* Line 298-303: Phase 2 - Use extracted StudentInfoCard component */}
-        <StudentInfoCard 
-          student={student}
-          studentPricing={studentPricing}
-          pricingLoading={pricingLoading}
-        />
-
-        {/* Line 305-317: Messages section */}
-        {successMessage && (
-          <div className="mb-4 p-3 bg-green-50 border border-green-200 rounded-lg">
-            <p className="text-green-800 text-sm">✅ {successMessage}</p>
-          </div>
-        )}
-
-        {error && (
-          <div className="mb-4 p-3 bg-red-50 border border-red-200 rounded-lg">
-            <p className="text-red-800 text-sm">❌ {error}</p>
-          </div>
-        )}
-
-        {/* Line 320-480: Form section */}
-        <div className="space-y-4">
-          {/* Quick select buttons */}
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-2">
-              Quick Select {studentPricing && `(${studentPricing.tier})`}
-            </label>
-            <div className="grid grid-cols-2 gap-3">
-              <button
-                type="button"
-                onClick={() => handleMembershipTypeChange("MONTHLY")}
-                disabled={loading || pricingLoading || isSubmitting}
-                className={`p-4 rounded-lg border text-sm transition-colors disabled:opacity-50 ${
-                  formData.membershipType === "MONTHLY"
-                    ? "bg-blue-50 border-blue-500 text-blue-700"
-                    : "bg-white border-gray-300 hover:bg-gray-50"
-                }`}
-              >
-                <div className="text-center">
-                  <div className="font-medium">Monthly</div>
-                  <div className="text-lg font-bold">
-                    ₱{membershipPrices.MONTHLY.toLocaleString()}
-                  </div>
-                </div>
-              </button>
-              <button
-                type="button"
-                onClick={() => handleMembershipTypeChange("YEARLY")}
-                disabled={loading || pricingLoading || isSubmitting}
-                className={`p-4 rounded-lg border text-sm transition-colors disabled:opacity-50 ${
-                  formData.membershipType === "YEARLY"
-                    ? "bg-blue-50 border-blue-500 text-blue-700"
-                    : "bg-white border-gray-300 hover:bg-gray-50"
-                }`}
-              >
-                <div className="text-center">
-                  <div className="font-medium">Yearly</div>
-                  <div className="text-lg font-bold">
-                    ₱{membershipPrices.YEARLY.toLocaleString()}
-                  </div>
-                </div>
-              </button>
-            </div>
-          </div>
-
-          {/* Amount input */}
-          <div>
-            <label htmlFor="amount" className="block text-sm font-medium text-gray-700 mb-1">
-              Payment Amount (₱)
-            </label>
-            <input
-              id="amount"
-              type="number"
-              name="amount"
-              value={formData.amount}
-              onChange={handleInputChange}
-              disabled={loading || pricingLoading || isSubmitting}
-              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 disabled:bg-gray-100"
-            />
-          </div>
-
-          {/* Payment method */}
-          <div>
-            <label htmlFor="method" className="block text-sm font-medium text-gray-700 mb-1">
-              Payment Method
-            </label>
-            <select
-              id="method"
-              name="method"
-              value={formData.method}
-              onChange={handleInputChange}
-              disabled={loading || isSubmitting}
-              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 disabled:bg-gray-100"
-            >
-              {paymentMethods.map((method) => (
-                <option key={method.value} value={method.value}>
-                  {method.label}
-                </option>
-              ))}
-            </select>
-          </div>
-
-          {/* Description */}
-          <div>
-            <label htmlFor="description" className="block text-sm font-medium text-gray-700 mb-1">
-              Description (Optional)
-            </label>
-            <textarea
-              id="description"
-              name="description"
-              value={formData.description}
-              onChange={handleInputChange}
-              rows="2"
-              disabled={loading || isSubmitting}
-              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 disabled:bg-gray-100"
-            />
-          </div>
-
-          {/* Extend membership checkbox */}
-          <div>
-            <label className="flex items-center space-x-2">
-              <input
-                type="checkbox"
-                name="extendMembership"
-                checked={formData.extendMembership}
-                onChange={handleInputChange}
-                disabled={loading || isSubmitting}
-                className="rounded text-blue-600"
-              />
-              <span className="text-sm font-medium text-gray-700">
-                Extend Membership ({formData.membershipType === "MONTHLY" ? "30 days" : "365 days"})
-              </span>
-            </label>
-          </div>
-
-          {/* Payment date picker */}
-          <div className="border border-blue-200 rounded-lg p-4 bg-blue-50">
-            <label className="block text-sm font-medium text-gray-700 mb-3">Payment Date</label>
-
-            <div className="space-y-3">
-              <label className="flex items-center space-x-2">
-                <input
-                  type="radio"
-                  name="paymentDateOption"
-                  value="today"
-                  checked={formData.paymentDateOption === "today"}
-                  onChange={(e) => handlePaymentDateOptionsChange(e.target.value)}
-                  disabled={loading || isSubmitting}
-                  className="text-blue-600"
-                />
-                <span className="text-sm font-medium">Today</span>
-              </label>
-              
-              <label className="flex items-center space-x-2">
-                <input
-                  type="radio"
-                  name="paymentDateOption"
-                  value="custom"
-                  checked={formData.paymentDateOption === "custom"}
-                  onChange={(e) => handlePaymentDateOptionsChange(e.target.value)}
-                  disabled={loading || isSubmitting}
-                  className="text-blue-600"
-                />
-                <span className="text-sm font-medium">Different Date</span>
-              </label>
-
-              {formData.paymentDateOption === "custom" && (
-                <div className="ml-6">
-                  <input
-                    type="date"
-                    value={formData.customPaymentDate}
-                    onChange={handleCustomDateChange}
-                    min={getMinDate()}
-                    max={getMaxDate()}
-                    disabled={loading || isSubmitting}
-                    className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 disabled:bg-gray-100"
-                  />
-                  {formData.customPaymentDate && (
-                    <div className="text-xs text-gray-600 mt-2">
-                      Selected: {getDisplayDate(formData.customPaymentDate)}
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-          </div>
-
-          {/* Action buttons */}
-          <div className="flex gap-3 pt-4">
-            <button
-              type="button"
-              onClick={handleClose}
-              className="flex-1 px-4 py-2 border border-gray-300 rounded-md text-gray-700 hover:bg-gray-50 transition-colors"
-              disabled={loading || isSubmitting}
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              onClick={handleSubmit}
-              disabled={
-                loading || 
-                pricingLoading || 
-                isSubmitting ||
-                !formData.amount || 
-                parseFloat(formData.amount) <= 0
-              }
-              className="flex-1 px-4 py-2 bg-green-600 text-white rounded-md hover:bg-green-700 disabled:opacity-50 disabled:cursor-not-allowed transition-colors font-medium"
-            >
-              {loading || isSubmitting ? "Processing..." : "Record Payment"}
-            </button>
-          </div>
-        </div>
-
-        {/* Line 487-504: Loading overlay */}
-        {(loading || pricingLoading || isSubmitting) && (
-          <div className="absolute inset-0 bg-white bg-opacity-90 flex items-center justify-center rounded-lg">
-            <div className="text-center">
-              <svg className="animate-spin h-8 w-8 text-blue-600 mx-auto mb-3" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
-                <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-              </svg>
-              <p className="text-sm text-gray-600">
-                {pricingLoading ? "Loading pricing..." : "Processing payment..."}
-              </p>
-            </div>
-          </div>
-        )}
+    <div className="mb-4 p-3 bg-blue-50 rounded-lg border border-blue-200">
+      {/* Line 28-35: Student basic info - name and email */}
+      <div className="flex items-center space-x-2 mb-1">
+        <span>👤</span>
+        <p className="font-medium text-blue-900">{student.name}</p>
       </div>
+      <p className="text-sm text-blue-700 ml-6">{student.email}</p>
+      
+      {/* Line 38-65: Pricing breakdown - only show when loaded */}
+      {studentPricing && !pricingLoading && (
+        <div className="mt-3 ml-6 p-3 bg-white rounded border">
+          <p className="text-sm font-medium text-gray-700 mb-2">Individual Rates</p>
+          
+          {/* Line 43-58: Monthly and yearly pricing grid */}
+          <div className="grid grid-cols-2 gap-3 text-sm">
+            {/* Monthly rate */}
+            <div className="text-center p-2 bg-blue-50 rounded">
+              <div className="font-semibold text-blue-900">
+                ₱{studentPricing.monthly?.toLocaleString()}
+              </div>
+              <div className="text-xs text-blue-600">Monthly</div>
+            </div>
+            
+            {/* Yearly rate */}
+            <div className="text-center p-2 bg-green-50 rounded">
+              <div className="font-semibold text-green-900">
+                ₱{studentPricing.yearly?.toLocaleString()}
+              </div>
+              <div className="text-xs text-green-600">Yearly</div>
+            </div>
+          </div>
+          
+          {/* Line 62-65: Grandfathered pricing indicator */}
+          {studentPricing.isLegacy && (
+            <p className="text-xs text-purple-600 mt-2">🌟 Grandfathered pricing</p>
+          )}
+        </div>
+      )}
     </div>
   );
 };
 
-export default PaymentModal;
+export default StudentInfoCard;
